FUNCTION When a record is created, add some values to specific cells to help with calculations and good record keeping
When a record is created, add some values to specific cells to help with calculations and good record keeping. This function is currently setup for use with a specific sheet for specific needs but it illustrates the possibilities. When glide apps add autocalculating components to run on sheet values some of this might not  be necessary.