NEEDS WORK - FUNCTION to change values in an item record to uncategorized if its associatied reference column value is ever deleted.
FUNCTION to change values in an item record to uncategorized if its associated reference column value is ever deleted. This way records no longer associated with an existing reference column group don't 'disappear' from a glideapp user's view and they can choose to recategorize it one that now exists.

Seems like it works when i manually edit the Interest sheet, but doesn't seem to be triggered when the Reference Sheet for languages is editted. Plus it seems like sometimes the array values get cached and it makes the value comparison based what was in the array before i deleted it from the reference list.